test(sim): wire spot sim genesis
