--- x/spot/module_simulation.go
@@ -1,33 +1,0 @@
-package spot
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/module"
-	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
-
-	"github.com/NibiruChain/nibiru/x/spot/simulation"
-	"github.com/NibiruChain/nibiru/x/spot/types"
-)
-
-// GenerateGenesisState creates a default GenState of the module
-func (AppModule) GenerateGenesisState(simState *module.SimulationState) {
-	accs := make([]string, len(simState.Accounts))
-	for i, acc := range simState.Accounts {
-		accs[i] = acc.Address.String()
-	}
-	dexGenesis := types.DefaultGenesis()
-	simState.GenState[types.ModuleName] = simState.Cdc.MustMarshalJSON(dexGenesis)
-}
-
-// ProposalContents doesn't return any content functions for governance proposals
-func (AppModule) ProposalContents(_ module.SimulationState) []simtypes.WeightedProposalMsg {
-	return nil
-}
-
-// RegisterStoreDecoder registers a decoder
-func (am AppModule) RegisterStoreDecoder(_ sdk.StoreDecoderRegistry) {}
-
-// WeightedOperations returns the all the spot module operations with their respective weights.
-func (am AppModule) WeightedOperations(_ module.SimulationState) []simtypes.WeightedOperation {
-	return simulation.WeightedOperations(am.accountKeeper, am.bankKeeper, am.keeper)
-}
